[c++,r] Move `nanoarrow` to `src/external`
