+++ apis/r/src/rinterface.cpp
@@ -1,7 +1,7 @@
-#include <Rcpp.h>               			// for R interface to C++
-#include <nanoarrow/r.h>        			// for C interface to Arrow (via R package)
-#include <tiledbsoma/utils/nanoarrow.hpp>   // for C/C++ interface to Arrow
-#include <RcppInt64>            			// for fromInteger64
+#include <Rcpp.h>                               // for R interface to C++
+#include <nanoarrow/r.h>                        // for C interface to Arrow (via R package)
+#include <tiledbsoma/nanoarrow/nanoarrow.hpp>   // for C/C++ interface to Arrow
+#include <RcppInt64>                            // for fromInteger64
 
 // we currently get deprecation warnings by default which are noisy
 #ifndef TILEDB_NO_API_DEPRECATION_WARNINGS
