+++ libtiledbsoma/CMakeLists.txt
@@ -4,7 +4,7 @@
 #
 # The MIT License
 #
-# Copyright (c) 2022-2023 TileDB, Inc.
+# Copyright (c) 2022-2024 TileDB, Inc.
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
