+++ libtiledbsoma/src/utils/arrow_adapter.h
@@ -8,7 +8,7 @@
 // https://arrow.apache.org/docs/format/Columnar.html#buffer-listing-for-each-layout
 // https://arrow.apache.org/docs/format/CDataInterface.html#exporting-a-simple-int32-array
 
-#include "nanoarrow.hpp"
+#include "tiledbsoma/nanoarrow/nanoarrow.hpp"
 
 namespace tiledbsoma {
